Move Transformers to its own package
