+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/parser2/ParserFactory.kt
@@ -19,7 +19,7 @@ package io.getstream.chat.android.client.parser2
 import io.getstream.chat.android.client.api2.mapping.DomainMapping
 import io.getstream.chat.android.client.api2.mapping.DtoMapping
 import io.getstream.chat.android.client.api2.mapping.EventMapping
-import io.getstream.chat.android.client.utils.ApiModelTransformers
+import io.getstream.chat.android.client.transformer.ApiModelTransformers
 
 internal object ParserFactory {
     fun createMoshiChatParser(
